Rename basePeerList -> base_list
