+++ elements/lisk-p2p/src/peer_directory/tried_peers.ts
@@ -20,7 +20,7 @@ import {
 
 import { P2PPeerInfo } from '../p2p_types';
 import { constructPeerIdFromPeerInfo } from '../utils';
-import { BasePeerList, CustomPeerInfo, PeerListConfig } from './basePeerList';
+import { BasePeerList, CustomPeerInfo, PeerListConfig } from './base_list';
 
 export interface TriedPeerConfig extends PeerListConfig {
 	readonly maxReconnectTries?: number;
